mod: compiler upgrade std c++20
- NO_INDEX define 추가
- c_variable class operator 수정
  rvalue 참조 연산사 && val 함수 추가
  '<, >, <=, >=, ==, !=' 연산자 const 통일
  VarFormat 함수 LPOLESTR 변수 추가
  c_string의 get_buffer() 함수 호출 명시적 변환
  value 관련 명시적 cast 변환

+++ Basic.cpp
@@ -803,7 +803,7 @@ STDMETHODIMP Basic::IsRoutine(/*[in]*/ BSTR str,
 							  /*[out, retval]*/ VARIANT_BOOL* bRet)
 {
 	c_string s_str = str;
-	*bRet = (m_engine.is_routine(s_str.get_buffer(), length) ? -1 : 0);
+	*bRet = (m_engine.is_routine(s_str.get_buffer(), static_cast<int>(length)) ? -1 : 0);
 	return S_OK;
 }
